requirements.txt
Django
djangorestframework
psycopg2>=2.8.6,<2.9
drf-spectacular